requirements.txt
click==7.1.2
Flask==1.1.2
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
click==7.1.2
Flask==1.1.2
Frozen-Flask==0.15
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyYAML==5.3.1
Werkzeug==1.0.1
